(SIO-2087) Start work on porting oioioi to six.
Ran modernize and made manual changes to preserve python2 stability.
You now can deploy and run oioioi in python3, although a lot more changes need
to be done in order to succesfully deploy it on py3, or even to pass tests.
This serves more as a non-breaking anything change that will allow further
work on this issue.

Change-Id: Ib63c209081c47a72fe15b10a553a9a2a433aec30