(SIO-2399) Black and isort bulk reformat
This commit contains a bulk reformat of all the files in oioioi
directory. The tools used for reformatting were black and isort. This
commit should be ignored from git blame. The config files for black and
isort will be included in a subsequent commit so that they are not
excluded from git blame.

Change-Id: Iac7b2e474534d1700234e798e73bd464ffe6f517